Allow use of http host and path during IPC
Previously the Host header was always `localhost` when connecting to a
Unix Domain Socket or a Windows Named Pipe. It was also not possible to
use any other path than `/`.

With this commit the URL is always treated as a URL and if you want to
supply a Unix Domain Socket or Windows Named Pipe, you instead have to
provide it using `--socketPath` (aliased `-S`).