fix: multipart form handling
Fix Content-Disposition header writing by exposing `form-data` `.append()` method options
in the `options` key of the form object when the type is `file`.
This allows passing a custom `filename` or `filepath` for a file. By default the `filename`
is set to the file name passed in the `path` key.

See:
* https://github.com/form-data/form-data/tree/v2.5.1#alternative-submission-methods
* https://github.com/form-data/form-data/tree/v2.5.1#void-append-string-field-mixed-value--mixed-options-