Enable SNI servername by default
Default to sending the SNI servername when the `servername` option is
not present or falsy. This behaviour matches browsers as well as popular
tools like `curl` or `openssl s_client`.

Because Node.js has 'deprecated' using an IP address for the servername,
do not infer it from the hostname if that is an IP address. Explitely
specifiying an IP address will still work.

Fixes: https://github.com/mcollina/autocannon/issues/233